# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PPIS API DOCUMENTATION",
        "description": "This is the Project Progres Information System API documentation.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/uploadTypes": {
            "get": {
                "tags": [
                    "uploadTypes"
                ],
                "summary": "Get a list of uploadTypess",
                "operationId": "68bcb348b3a75dd926d8f9412eaa2806",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "headers": {
                            "Cache-Control": {
                                "description": "Cache control header",
                                "schema": {
                                    "type": "string",
                                    "example": "no-cache, private"
                                }
                            },
                            "Content-Type": {
                                "description": "Content type header",
                                "schema": {
                                    "type": "string",
                                    "example": "application/json; charset=UTF-8"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "upload_type_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "upload_name": {
                                                        "type": "string",
                                                        "example": "ROLE NATIONAL"
                                                    },
                                                    "created_by": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Mohammed"
                                                    },
                                                    "middle_name": {
                                                        "type": "string",
                                                        "example": "Abdalla"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Bakar"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-28 11:30:25"
                                                    },
                                                    "deleted_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-28 11:30:25"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-28 11:30:25"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "uploadTypes"
                ],
                "summary": "Store a new uploadTypes",
                "operationId": "69f6068109b88a41aae2bc4bc0f94a58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "upload_name": {
                                        "type": "string"
                                    },
                                    "upload_excel": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "headers": {
                            "Cache-Control": {
                                "description": "Cache control header",
                                "schema": {
                                    "type": "string",
                                    "example": "no-cache, private"
                                }
                            },
                            "Content-Type": {
                                "description": "Content type header",
                                "schema": {
                                    "type": "string",
                                    "example": "application/json; charset=UTF-8"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "statusCode": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/uploadTypes/{upload_type_id}": {
            "get": {
                "tags": [
                    "uploadTypes"
                ],
                "summary": "Get a specific uploadTypes",
                "operationId": "fb7d9d576babcc3fa03f7eb5efdda286",
                "parameters": [
                    {
                        "name": "upload_type_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "headers": {
                            "Cache-Control": {
                                "description": "Cache control header",
                                "schema": {
                                    "type": "string",
                                    "example": "no-cache, private"
                                }
                            },
                            "Content-Type": {
                                "description": "Content type header",
                                "schema": {
                                    "type": "string",
                                    "example": "application/json; charset=UTF-8"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "upload_type_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "upload_name": {
                                                        "type": "string",
                                                        "example": "ROLE NATIONAL"
                                                    },
                                                    "created_by": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Mohammed"
                                                    },
                                                    "middle_name": {
                                                        "type": "string",
                                                        "example": "Abdalla"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Bakar"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-28 11:30:25"
                                                    },
                                                    "deleted_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-28 11:30:25"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-28 11:30:25"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "uploadTypes"
                ],
                "summary": "Update a uploadTypes",
                "operationId": "3c7b33a91bc8b97f0edbd3edbe70cea1",
                "parameters": [
                    {
                        "name": "upload_type_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "upload_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "headers": {
                            "Cache-Control": {
                                "description": "Cache control header",
                                "schema": {
                                    "type": "string",
                                    "example": "no-cache, private"
                                }
                            },
                            "Content-Type": {
                                "description": "Content type header",
                                "schema": {
                                    "type": "string",
                                    "example": "application/json; charset=UTF-8"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "statusCode": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}